API_calories/requirements.txt
2to3==1.0
albumentations==1.3.1
altair==5.0.1
annotated-types==0.5.0
anyio==3.7.1
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
caer==2.0.8
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comet-ml==3.33.5
comm==0.1.3
configobj==5.0.8
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
dulwich==0.21.5
everett==3.1.0
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.100.0
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.40.0
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
idna==3.4
imageio==2.31.1
importlib-metadata==6.7.0
ipykernel==6.23.2
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.1
keras==2.13.1
kiwisolver==1.4.4
lazy_loader==0.3
libclang==16.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
mypy==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.7.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pydantic==2.0.2
pydantic_core==2.1.2
pydeck==0.8.1b0
pygame==2.1.0
Pygments==2.15.1
pylance==0.4.21
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-box==6.1.0
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0
pyzmq==25.1.0
qudida==0.0.4
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.4.2
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
semantic-version==2.10.0
sentry-sdk==1.26.0
simplejson==3.19.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
starlette==0.27.0
streamlit==1.24.0
streamlit-embedcode==0.1.2
sympy==1.12
tenacity==8.2.2
tensorboard-data-server==0.7.1
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.7.4
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3.1
ultralytics==8.0.121
urllib3==1.26.16
uvicorn==0.22.0
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
websocket-client==1.3.3
Werkzeug==2.3.6
wrapt==1.14.1
wurlitzer==3.0.3
zipp==3.15.0
